Converted remaining charts to use a custom component instead of vue-chartjs. Re-enabled the legends and tooltips in the charts (had to register them due to ChartJS's tree shaking). Fixed type errors, waiting on a chartjs-plugin-zoom update to remove errors in their type stubs.
