+++ package.json
@@ -20,7 +20,6 @@
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.33",
     "vue": "^2.6.12",
-    "vue-chartjs": "^3.5.1",
     "vue-class-component": "^7.2.6",
     "vue-property-decorator": "^8.5.1",
     "vue-router": "^3.5.1",
