+++ src/components/project_admin/project_stats/project_stats.vue
@@ -194,7 +194,7 @@ import { handle_api_errors_async, handle_global_errors_async } from '@/error_han
 import { SafeMap } from '@/safe_map';
 import { toggle } from '@/utils';
 
-import FirstSubmissionTimeVsFinalScore from './first_submission_time_vs_final_score';
+import FirstSubmissionTimeVsFinalScore from './first_submission_time_vs_final_score.vue';
 import PassFailCounts from './pass_fail_counts.vue';
 import SubmissionScoreHistogram from './submission_score_histogram/submission_score_histogram.vue';
 import SubmissionsOverTimeGraph from './submissions_over_time_graph.vue';
